refactor(ui): move @Preview goldens into sibling *Previews.kt files
The screenshot-golden previews lived in the same file as the composables they
exercise, so a production file carried a large tail of preview/test scaffolding.
Extract them into sibling files in the same (main) source set - the preview pane
and Roborazzi's generator both still scan them - as an exemplar for doing this
repo-wide:

  ui/common/CoverImage.kt      -> ui/common/CoverImagePreviews.kt
  ui/library/LibraryScreen.kt  -> ui/library/LibraryScreenPreviews.kt

Kotlin has no friend/package-private visibility and `private` is file-scoped, so
a preview in another file can only reach a composable that is at least internal.
Kept that to the minimum:

  - CoverImage: only LoadingTileContent became internal (there is no public way
    to freeze the spinner-shown state). The no-cover preview routes through the
    public CoverImage(coverUrl = null), so NoCoverTile stays private.
  - LibraryScreen: only LibraryContent became internal.

Goldens are byte-identical - git records all 21 as renames - because only the
generated test class name changed (CoverImageKt -> CoverImagePreviewsKt,
LibraryScreenKt -> LibraryScreenPreviewsKt). No pixels moved.

Refs #97
