+++ app/src/main/java/io/github/xexanos/ratatoskr/ui/common/CoverImage.kt
@@ -5,29 +5,20 @@
  */
 package io.github.xexanos.ratatoskr.ui.common
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.sizeIn
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Canvas
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.graphics.Paint
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.semantics.clearAndSetSemantics
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import coil3.ImageLoader
@@ -40,7 +31,6 @@ import io.github.xexanos.ratatoskr.ui.KnotMark
 import io.github.xexanos.ratatoskr.ui.UiTestTags
 import io.github.xexanos.ratatoskr.ui.rememberDelayedVisible
 import io.github.xexanos.ratatoskr.ui.theme.LocalReducedMotion
-import io.github.xexanos.ratatoskr.ui.theme.RatatoskrTheme
 
 /**
  * The cover [ImageLoader], provided by the activity from the AppContainer. A CompositionLocal
@@ -135,8 +125,11 @@ private fun LoadingTile(withSpinner: Boolean) {
 // The visible layout of the loading tile, split from the delay gate so the previews (and any
 // screenshot golden) can render the spinner-shown state directly - rememberDelayedVisible never
 // elapses in a static render, so a preview of LoadingTile would always be the bare tile.
+// Internal, not private, for exactly that: the sibling CoverImagePreviews.kt drives it, and no
+// public seam reaches this spinner-forced state (a stuck loader under a paused clock shows the
+// bare tile). The no-cover preview needs no such hook - it goes through CoverImage(null).
 @Composable
-private fun LoadingTileContent(showSpinner: Boolean) {
+internal fun LoadingTileContent(showSpinner: Boolean) {
     Box(
         contentAlignment = Alignment.Center,
         modifier = Modifier.fillMaxSize().testTag(UiTestTags.COVER_LOADING),
@@ -175,129 +168,3 @@ private fun NoCoverTile() {
         )
     }
 }
-
-// --- Previews / screenshot goldens ---------------------------------------------------------
-//
-// Both placeholder tiles at both sizes they ship at - the 56 dp list-row thumbnail and the
-// 260 dp now-playing tile - in light and dark. These render without a server or ImageLoader
-// (the coverUrl != null path needs one; the tiles do not), so Roborazzi turns each @Preview
-// into a screenshot golden (build.gradle.kts: generateComposePreviewRobolectricTests): drift
-// in the knot mark, the spinner size, or the tile tones is caught pixel-for-pixel. The loading
-// previews force the spinner visible via LoadingTileContent - the delayed gate never elapses
-// in a static render.
-
-private const val ROW_TILE_DP = 56 // list-row / shelf thumbnail
-private const val NOW_PLAYING_TILE_DP = 260 // now-playing cover
-
-// The production framing: the tonal Surface CoverImage wraps every tile in, sized to one tile.
-@Composable
-private fun CoverTilePreview(dark: Boolean, sizeDp: Int, content: @Composable () -> Unit) {
-    RatatoskrTheme(darkTheme = dark) {
-        Surface(
-            shape = MaterialTheme.shapes.small,
-            color = MaterialTheme.colorScheme.secondaryContainer,
-            modifier = Modifier.size(sizeDp.dp),
-        ) {
-            content()
-        }
-    }
-}
-
-@Preview(name = "No-cover 56dp light")
-@Composable
-internal fun NoCoverTile56LightPreview() =
-    CoverTilePreview(dark = false, sizeDp = ROW_TILE_DP) { NoCoverTile() }
-
-@Preview(name = "No-cover 56dp dark")
-@Composable
-internal fun NoCoverTile56DarkPreview() =
-    CoverTilePreview(dark = true, sizeDp = ROW_TILE_DP) { NoCoverTile() }
-
-@Preview(name = "No-cover 260dp light")
-@Composable
-internal fun NoCoverTile260LightPreview() =
-    CoverTilePreview(dark = false, sizeDp = NOW_PLAYING_TILE_DP) { NoCoverTile() }
-
-@Preview(name = "No-cover 260dp dark")
-@Composable
-internal fun NoCoverTile260DarkPreview() =
-    CoverTilePreview(dark = true, sizeDp = NOW_PLAYING_TILE_DP) { NoCoverTile() }
-
-@Preview(name = "Loading 56dp light")
-@Composable
-internal fun LoadingTile56LightPreview() =
-    CoverTilePreview(dark = false, sizeDp = ROW_TILE_DP) { LoadingTileContent(showSpinner = true) }
-
-@Preview(name = "Loading 56dp dark")
-@Composable
-internal fun LoadingTile56DarkPreview() =
-    CoverTilePreview(dark = true, sizeDp = ROW_TILE_DP) { LoadingTileContent(showSpinner = true) }
-
-@Preview(name = "Loading 260dp light")
-@Composable
-internal fun LoadingTile260LightPreview() =
-    CoverTilePreview(dark = false, sizeDp = NOW_PLAYING_TILE_DP) { LoadingTileContent(showSpinner = true) }
-
-@Preview(name = "Loading 260dp dark")
-@Composable
-internal fun LoadingTile260DarkPreview() =
-    CoverTilePreview(dark = true, sizeDp = NOW_PLAYING_TILE_DP) { LoadingTileContent(showSpinner = true) }
-
-// Goldens for the Fit treatment (KDoc above) at row size, in the two aspect ratios: portrait
-// 2:3 and square. The portrait golden must show tonal bands to the sides and a round disc; the
-// square one must fill edge to edge. These render an Image directly rather than CoverImage: the
-// production success path is a SubcomposeAsyncImage, whose load never resolves under Roborazzi's
-// paused composition clock (it would only ever golden the loading tile), so a synthetic bitmap
-// through the same Surface + ContentScale.Fit stands in for the decoded cover - the same spirit
-// as LoadingTileContent bypassing the delayed spinner gate.
-
-// A stand-in cover: a solid fill with a centered disc, so any non-uniform scaling shows as an
-// out-of-round disc. The two colors come from the theme (not literals - the UX gate forbids
-// hardcoded Color(0x...), scripts/check-ux.sh); they only need to contrast with the tile's
-// secondaryContainer band for the letterbox to read.
-private fun sampleCoverBitmap(widthPx: Int, heightPx: Int, fill: Color, mark: Color): ImageBitmap {
-    val bitmap = ImageBitmap(widthPx, heightPx)
-    val canvas = Canvas(bitmap)
-    val paint = Paint().apply { color = fill }
-    canvas.drawRect(0f, 0f, widthPx.toFloat(), heightPx.toFloat(), paint)
-    paint.color = mark
-    canvas.drawCircle(Offset(widthPx / 2f, heightPx / 2f), minOf(widthPx, heightPx) / 3f, paint)
-    return bitmap
-}
-
-@Composable
-private fun SampleCoverTile(dark: Boolean, widthPx: Int, heightPx: Int) {
-    CoverTilePreview(dark = dark, sizeDp = ROW_TILE_DP) {
-        val fill = MaterialTheme.colorScheme.primary
-        val mark = MaterialTheme.colorScheme.onPrimary
-        val bitmap = remember(widthPx, heightPx, fill, mark) {
-            sampleCoverBitmap(widthPx, heightPx, fill, mark)
-        }
-        Image(
-            bitmap = bitmap,
-            contentDescription = null,
-            contentScale = ContentScale.Fit,
-            modifier = Modifier.fillMaxSize(),
-        )
-    }
-}
-
-@Preview(name = "Cover portrait 56dp light")
-@Composable
-internal fun CoverPortrait56LightPreview() =
-    SampleCoverTile(dark = false, widthPx = 40, heightPx = 60)
-
-@Preview(name = "Cover portrait 56dp dark")
-@Composable
-internal fun CoverPortrait56DarkPreview() =
-    SampleCoverTile(dark = true, widthPx = 40, heightPx = 60)
-
-@Preview(name = "Cover square 56dp light")
-@Composable
-internal fun CoverSquare56LightPreview() =
-    SampleCoverTile(dark = false, widthPx = 60, heightPx = 60)
-
-@Preview(name = "Cover square 56dp dark")
-@Composable
-internal fun CoverSquare56DarkPreview() =
-    SampleCoverTile(dark = true, widthPx = 60, heightPx = 60)
