+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.CoverPortrait56DarkPreview.Cover_portrait_56dp_dark.png
