+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.CoverPortrait56LightPreview.Cover_portrait_56dp_light.png
