+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.CoverSquare56DarkPreview.Cover_square_56dp_dark.png
