+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.CoverSquare56LightPreview.Cover_square_56dp_light.png
