+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.LoadingTile260DarkPreview.Loading_260dp_dark.png
