+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.LoadingTile260LightPreview.Loading_260dp_light.png
