+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.LoadingTile56DarkPreview.Loading_56dp_dark.png
