+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.LoadingTile56LightPreview.Loading_56dp_light.png
