+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.NoCoverTile260DarkPreview.No-cover_260dp_dark.png
