+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.NoCoverTile260LightPreview.No-cover_260dp_light.png
