+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.NoCoverTile56DarkPreview.No-cover_56dp_dark.png
