+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.common.CoverImagePreviewsKt.NoCoverTile56LightPreview.No-cover_56dp_light.png
