+++ app/src/test/screenshots/io.github.xexanos.ratatoskr.ui.library.LibraryScreenPreviewsKt.LibrarySearchingShelfHiddenPreview.Library_-_searching,_shelf_hidden_W360dp_H800dp.png
